updated to handle different calls in different fragments instead of passing enum
