+++ app/src/main/java/com/example/recipeapp/Fragments/BookmarkFeedFragment.java
@@ -16,6 +16,7 @@ import android.view.ViewGroup;
 import com.example.recipeapp.Adapters.RecipesAdapter;
 import com.example.recipeapp.Models.Recipe;
 import com.example.recipeapp.R;
+import com.example.recipeapp.Room.RecipesRepository;
 import com.example.recipeapp.viewmodels.FeedViewModel;
 
 import java.util.List;
@@ -25,7 +26,7 @@ public class BookmarkFeedFragment extends FeedFragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        viewModel.fetch(FeedViewModel.DataSource.LOCAL_SQL_DB, null);
+        viewModel.fetch(RecipesRepository.DataSource.LOCAL_SQL_DB, null);
         viewModel.allRecipes.observe(getViewLifecycleOwner(), new Observer<List<Recipe>>() {
             @Override
             public void onChanged(List<Recipe> recipes) {
