+++ app/src/main/java/com/example/recipeapp/Fragments/SearchFeedFragment.java
@@ -18,18 +18,29 @@ import com.example.recipeapp.Adapters.RecipesAdapter;
 import com.example.recipeapp.Models.API.ApiCallParams;
 import com.example.recipeapp.Models.Recipe;
 import com.example.recipeapp.R;
+import com.example.recipeapp.Room.RecipesRepository;
 import com.example.recipeapp.viewmodels.FeedViewModel;
 
 import java.util.List;
 
 public class SearchFeedFragment extends FeedFragment {
 
+    private ApiCallParams params;
 
+    public SearchFeedFragment() {
+        // Required empty public constructor
+    }
+
+    public static SearchFeedFragment newInstance(ApiCallParams params) {
+        SearchFeedFragment fragment = new SearchFeedFragment();
+        fragment.params = params;
+        return fragment;
+    }
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        viewModel.fetch(FeedViewModel.DataSource.LOCAL_SQL_DB, null);
+        viewModel.fetch(RecipesRepository.DataSource.API_CALL, params);
         viewModel.allRecipes.observe(getViewLifecycleOwner(), new Observer<List<Recipe>>() {
             @Override
             public void onChanged(List<Recipe> recipes) {
