+++ app/src/main/java/com/example/recipeapp/Fragments/SearchFragment.java
@@ -142,8 +142,7 @@ public class SearchFragment extends Fragment {
                 putWithEmptyCheck(String.join(",", included)), putWithEmptyCheck(String.join(",", excluded)),
                 putWithEmptyCheck(binding.spinnerType.getSelectedItem()), putWithEmptyCheck(binding.edTime.getText().toString()));
 
-        getActivity().getSupportFragmentManager().beginTransaction().replace(R.id.fragmentPlaceholder, new SearchFeedFragment()).commit();
-        ((MainActivity)(getActivity())).getBinding().bottomNavigation.setSelectedItemId(R.id.feed);
+        getActivity().getSupportFragmentManager().beginTransaction().replace(R.id.fragmentPlaceholder, SearchFeedFragment.newInstance(params)).commit();
     }
 
     private String putWithEmptyCheck(CharSequence chars) {
